🎨 Reminderコマンドのリファクタリング
